requirements.txt
annotated-types==0.6.0
anyio==4.2.0
appdirs==1.4.4
certifi==2023.11.17
charset-normalizer==3.3.2
distro==1.9.0
geographiclib==2.0
geopy==2.4.1
google_search_results==2.4.2
h11==0.14.0
httpcore==1.0.2
httpx==0.26.0
idna==3.6
importlib-metadata==7.0.1
openai==1.6.1
pexels-api==1.0.1
pydantic==2.5.3
pydantic_core==2.14.6
pyee==8.2.2
pyppeteer==1.0.2
python-dotenv==1.0.0
requests==2.31.0
sniffio==1.3.0
tqdm==4.66.1
typing_extensions==4.9.0
urllib3==1.26.18
websockets==10.4
zipp==3.17.0
